backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
redis==5.0.4
minio==7.2.7
python-multipart==0.0.9
python-dotenv==1.0.1
httpx==0.27.0
openai==1.30.0
langfuse==2.26.1
pymilvus==2.4.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
